Разделяет на моки и константы
