+++ src/components/private-route.tsx
@@ -1,4 +1,4 @@
-import { AuthorizationStatus} from '../const';
+import { AuthorizationStatus} from '../mocks';
 import { AppRoute } from './app';
 import { Navigate } from 'react-router-dom';
