+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/app';
-import mocks from './const';
+import mocks from './mocks';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
