+++ src/mocks.ts
@@ -1,0 +1,20 @@
+const mocks = [{
+  'id':'e3274362-f06e-4870-ab15-a08b9163ea63',
+  'title':'Canal View Prinsengracht',
+  'type':'hotel',
+  'price':203,
+  'previewImage':'https://13.design.pages.academy/static/hotel/12.jpg','city':{
+    'name':'Paris',
+    'location':{
+      'latitude':48.85661,
+      'longitude':2.351499,
+      'zoom':13}},
+  'location':{
+    'latitude':48.868610000000004,
+    'longitude':2.342499,'zoom':16},
+  'isFavorite':true,
+  'isPremium':false,
+  'rating':2.6},
+{'id':'6de1df75-3659-4aad-8d09-c076d19b49fe','title':'The house among olive ','type':'apartment','price':348,'previewImage':'https://13.design.pages.academy/static/hotel/17.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.858610000000006,'longitude':2.330499,'zoom':16},'isFavorite':false,'isPremium':false,'rating':2.8},{'id':'e360a8c2-9f6e-464d-afa5-74c5bb5b5c84','title':'Amazing and Extremely Central Flat','type':'room','price':174,'previewImage':'https://13.design.pages.academy/static/hotel/6.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.834610000000005,'longitude':2.335499,'zoom':16},'isFavorite':false,'isPremium':false,'rating':3.2},{'id':'257d4584-70df-4f9f-9570-3aafe154ca10','title':'The house among olive ','type':'apartment','price':447,'previewImage':'https://13.design.pages.academy/static/hotel/14.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.85761,'longitude':2.358499,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2.3},{'id':'322ecbb8-60c0-4d8b-a8af-905280248006','title':'The Pondhouse - A Magical Place','type':'house','price':502,'previewImage':'https://13.design.pages.academy/static/hotel/14.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.87561,'longitude':2.375499,'zoom':16},'isFavorite':false,'isPremium':true,'rating':3.6},{'id':'55629d5f-c51e-4834-9d6b-2000c0e8a23f','title':'Canal View Prinsengracht','type':'room','price':289,'previewImage':'https://13.design.pages.academy/static/hotel/12.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.87961000000001,'longitude':2.353499,'zoom':16},'isFavorite':false,'isPremium':false,'rating':2.6},{'id':'0cc32d4e-2714-4295-9e1e-a6510f0cd49a','title':'Loft Studio in the Central Area','type':'room','price':289,'previewImage':'https://13.design.pages.academy/static/hotel/1.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.834610000000005,'longitude':2.364499,'zoom':16},'isFavorite':false,'isPremium':true,'rating':4.6},{'id':'25d92cfe-a2b8-4b8c-a75e-b3aa1113effc','title':'Penthouse, 4-5 rooms + 5 balconies','type':'hotel','price':417,'previewImage':'https://13.design.pages.academy/static/hotel/9.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.837610000000005,'longitude':2.3454990000000002,'zoom':16},'isFavorite':false,'isPremium':false,'rating':3.1},{'id':'8cca26b1-1ab6-4e0a-b1ee-50cf6d2e913d','title':'Tile House','type':'apartment','price':432,'previewImage':'https://13.design.pages.academy/static/hotel/2.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.84761,'longitude':2.356499,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1.5},{'id':'98dd3f31-e91e-452c-90f8-4f4078e324ba','title':'The Pondhouse - A Magical Place','type':'room','price':101,'previewImage':'https://13.design.pages.academy/static/hotel/7.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.862610000000004,'longitude':2.369499,'zoom':16},'isFavorite':false,'isPremium':false,'rating':2.5},{'id':'2df2288c-490b-442b-9e29-dda3db6b11fb','title':'Beautiful & luxurious apartment at great location','type':'apartment','price':225,'previewImage':'https://13.design.pages.academy/static/hotel/1.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.83861,'longitude':2.350499,'zoom':16},'isFavorite':false,'isPremium':true,'rating':4.8},{'id':'cbae3688-f5b9-443f-aa70-88a0439ae261','title':'Waterfront with extraordinary view','type':'apartment','price':182,'previewImage':'https://13.design.pages.academy/static/hotel/14.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.861610000000006,'longitude':2.340499,'zoom':16},'isFavorite':true,'isPremium':true,'rating':3.9},{'id':'b66e62ca-d63a-4e3a-a7e5-fe274c77a66b','title':'The Pondhouse - A Magical Place','type':'house','price':976,'previewImage':'https://13.design.pages.academy/static/hotel/20.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.87861,'longitude':2.357499,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1.7},{'id':'3a5beeff-8c78-4037-9f1c-aa22378e78d1','title':'Loft Studio in the Central Area','type':'apartment','price':473,'previewImage':'https://13.design.pages.academy/static/hotel/2.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.877610000000004,'longitude':2.333499,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2},{'id':'73417422-7fe6-4a51-aa3d-34e0029132d3','title':'Tile House','type':'house','price':361,'previewImage':'https://13.design.pages.academy/static/hotel/16.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.83961,'longitude':2.342499,'zoom':16},'isFavorite':false,'isPremium':false,'rating':3.3},{'id':'b6caa299-58af-454b-9f95-ce34764676fb','title':'Beautiful & luxurious apartment at great location','type':'house','price':418,'previewImage':'https://13.design.pages.academy/static/hotel/6.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.865610000000004,'longitude':2.350499,'zoom':16},'isFavorite':false,'isPremium':false,'rating':4.5},{'id':'bae70fc3-68ac-45a9-8dfc-be0a4970a820','title':'Canal View Prinsengracht','type':'hotel','price':162,'previewImage':'https://13.design.pages.academy/static/hotel/4.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.846610000000005,'longitude':2.374499,'zoom':16},'isFavorite':false,'isPremium':false,'rating':3},{'id':'ba756105-5535-4872-801e-a11ba376734a','title':'Amazing and Extremely Central Flat','type':'house','price':633,'previewImage':'https://13.design.pages.academy/static/hotel/6.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.837610000000005,'longitude':2.364499,'zoom':16},'isFavorite':false,'isPremium':false,'rating':4.8},{'id':'99ab2689-ff66-4445-8355-deac91994cf1','title':'Nice, cozy, warm big bed apartment','type':'hotel','price':378,'previewImage':'https://13.design.pages.academy/static/hotel/10.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.843610000000005,'longitude':2.338499,'zoom':16},'isFavorite':false,'isPremium':false,'rating':4.5},{'id':'aa0b52ae-0d40-4f3b-9c51-139340699656','title':'Loft Studio in the Central Area','type':'hotel','price':377,'previewImage':'https://13.design.pages.academy/static/hotel/3.jpg','city':{'name':'Paris','location':{'latitude':48.85661,'longitude':2.351499,'zoom':13}},'location':{'latitude':48.84461,'longitude':2.374499,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1.1},{'id':'cf1f1e9b-2c70-46cc-8c9f-fe7cfe63b135','title':'Beautiful & luxurious apartment at great location','type':'house','price':157,'previewImage':'https://13.design.pages.academy/static/hotel/11.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.950361,'longitude':6.961974,'zoom':16},'isFavorite':false,'isPremium':true,'rating':4.7},{'id':'37b68097-3f89-4a73-b4c9-7e6021c184ee','title':'The Pondhouse - A Magical Place','type':'apartment','price':439,'previewImage':'https://13.design.pages.academy/static/hotel/8.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.932361,'longitude':6.937974,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2.4},{'id':'ea1a307d-f32f-4522-ac56-43e7ee44a7dc','title':'Penthouse, 4-5 rooms + 5 balconies','type':'house','price':457,'previewImage':'https://13.design.pages.academy/static/hotel/4.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.934361,'longitude':6.943974,'zoom':16},'isFavorite':false,'isPremium':true,'rating':4.4},{'id':'e08a1d48-e362-4651-bbf5-d326f6350df8','title':'Waterfront with extraordinary view','type':'hotel','price':488,'previewImage':'https://13.design.pages.academy/static/hotel/10.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.947361,'longitude':6.9799739999999995,'zoom':16},'isFavorite':false,'isPremium':false,'rating':2.7},{'id':'3e97f18b-8355-47e0-9fc0-50969fa4d3c7','title':'Wood and stone place','type':'apartment','price':331,'previewImage':'https://13.design.pages.academy/static/hotel/18.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.960361,'longitude':6.967974,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1.8},{'id':'8fea2e4d-d503-4ee0-aee5-124f5166eef9','title':'Loft Studio in the Central Area','type':'hotel','price':430,'previewImage':'https://13.design.pages.academy/static/hotel/10.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.941361,'longitude':6.956974,'zoom':16},'isFavorite':false,'isPremium':false,'rating':4.1},{'id':'b264833d-8af9-4bd7-9457-8d5225e8daef','title':'Loft Studio in the Central Area','type':'room','price':273,'previewImage':'https://13.design.pages.academy/static/hotel/3.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.916361,'longitude':6.944974,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1.8},{'id':'ab09dda0-7ba5-4015-9d3c-a880ef3ea656','title':'Canal View Prinsengracht','type':'apartment','price':418,'previewImage':'https://13.design.pages.academy/static/hotel/20.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.949361,'longitude':6.976974,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2},{'id':'dd34b49f-5054-4670-823b-c737e38d4a3f','title':'Nice, cozy, warm big bed apartment','type':'house','price':536,'previewImage':'https://13.design.pages.academy/static/hotel/17.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.913361,'longitude':6.9509739999999995,'zoom':16},'isFavorite':false,'isPremium':false,'rating':2.1},{'id':'655ff7ba-97fb-4edc-b2e5-227c5f2ef35b','title':'The house among olive ','type':'apartment','price':324,'previewImage':'https://13.design.pages.academy/static/hotel/19.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.930361,'longitude':6.937974,'zoom':16},'isFavorite':false,'isPremium':true,'rating':3.2},{'id':'28d93dd7-c5a4-444e-9ca5-7ec0c4cef900','title':'House in countryside','type':'house','price':310,'previewImage':'https://13.design.pages.academy/static/hotel/11.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.960361,'longitude':6.9509739999999995,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1.9},{'id':'e34df19a-3134-4847-948c-86c54e803327','title':'Canal View Prinsengracht','type':'room','price':238,'previewImage':'https://13.design.pages.academy/static/hotel/18.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.918461,'longitude':6.969974,'zoom':16},'isFavorite':false,'isPremium':false,'rating':1.1},{'id':'8abf52ab-7eb1-441e-be4e-e03b7de0d8fb','title':'Nice, cozy, warm big bed apartment','type':'room','price':225,'previewImage':'https://13.design.pages.academy/static/hotel/17.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.957361,'longitude':6.9509739999999995,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2.6},{'id':'4fc2325c-b520-4aea-bd18-8fe92d12513b','title':'Nice, cozy, warm big bed apartment','type':'apartment','price':470,'previewImage':'https://13.design.pages.academy/static/hotel/7.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.951361,'longitude':6.944974,'zoom':16},'isFavorite':false,'isPremium':true,'rating':3.2},{'id':'2225e147-bfd8-4d17-b91a-64a7e782903b','title':'The house among olive ','type':'apartment','price':293,'previewImage':'https://13.design.pages.academy/static/hotel/14.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.959361,'longitude':6.978974,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1.8},{'id':'88cc9e08-590d-4117-ad4d-d9d0059f694a','title':'Beautiful & luxurious apartment at great location','type':'room','price':216,'previewImage':'https://13.design.pages.academy/static/hotel/14.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.932361,'longitude':6.960974,'zoom':16},'isFavorite':false,'isPremium':false,'rating':3.3},{'id':'54772098-54ee-47ba-b443-7e3ed63f9fa1','title':'Loft Studio in the Central Area','type':'apartment','price':267,'previewImage':'https://13.design.pages.academy/static/hotel/12.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.954361,'longitude':6.982974,'zoom':16},'isFavorite':false,'isPremium':false,'rating':3.1},{'id':'c86635d4-b983-41ba-9b3f-75c33835a8b7','title':'Nice, cozy, warm big bed apartment','type':'apartment','price':368,'previewImage':'https://13.design.pages.academy/static/hotel/8.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.945361,'longitude':6.962974,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2.3},{'id':'312746e3-98ae-4814-93fb-ad435d43371b','title':'Nice, cozy, warm big bed apartment','type':'hotel','price':209,'previewImage':'https://13.design.pages.academy/static/hotel/2.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.917361,'longitude':6.977974,'zoom':16},'isFavorite':false,'isPremium':false,'rating':3.6},{'id':'d5781111-2c5c-440e-8e6c-a04ddd1ff08f','title':'Beautiful & luxurious apartment at great location','type':'apartment','price':100,'previewImage':'https://13.design.pages.academy/static/hotel/3.jpg','city':{'name':'Cologne','location':{'latitude':50.938361,'longitude':6.959974,'zoom':13}},'location':{'latitude':50.945361,'longitude':6.935974,'zoom':16},'isFavorite':false,'isPremium':true,'rating':3.4},{'id':'897c04f1-b349-4dde-bbc7-3cc1ef3492ea','title':'Perfectly located Castro','type':'apartment','price':499,'previewImage':'https://13.design.pages.academy/static/hotel/12.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.854557,'longitude':4.364697,'zoom':16},'isFavorite':false,'isPremium':false,'rating':2.6},{'id':'7a3e8bff-33fe-4b1a-bc9e-27445bdb7671','title':'Amazing and Extremely Central Flat','type':'hotel','price':332,'previewImage':'https://13.design.pages.academy/static/hotel/7.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.867557,'longitude':4.371696999999999,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1.3},{'id':'bdc711aa-37a8-4910-86c7-723453c3df8d','title':'The Pondhouse - A Magical Place','type':'room','price':278,'previewImage':'https://13.design.pages.academy/static/hotel/3.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.827557,'longitude':4.336697,'zoom':16},'isFavorite':false,'isPremium':false,'rating':3.7},{'id':'12ac8428-225f-472a-8c6c-4146f17e5818','title':'Waterfront with extraordinary view','type':'house','price':647,'previewImage':'https://13.design.pages.academy/static/hotel/2.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.833557,'longitude':4.374696999999999,'zoom':16},'isFavorite':false,'isPremium':false,'rating':4.7},{'id':'3161ba7b-e085-46fd-8381-f22dbe29801e','title':'Nice, cozy, warm big bed apartment','type':'house','price':221,'previewImage':'https://13.design.pages.academy/static/hotel/11.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.837557,'longitude':4.339697,'zoom':16},'isFavorite':false,'isPremium':false,'rating':4.5},{'id':'99cd72b5-60b7-4e1e-9374-2f6c1fcf7411','title':'The house among olive ','type':'apartment','price':273,'previewImage':'https://13.design.pages.academy/static/hotel/6.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.849557,'longitude':4.374696999999999,'zoom':16},'isFavorite':false,'isPremium':false,'rating':4.7},{'id':'e0cedbed-8631-4d99-b0d8-6ec8b46159f1','title':'Amazing and Extremely Central Flat','type':'house','price':785,'previewImage':'https://13.design.pages.academy/static/hotel/16.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.852557,'longitude':4.3376969999999995,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2.2},{'id':'84483746-fa07-4b7c-9bba-74441ea7c217','title':'Perfectly located Castro','type':'room','price':241,'previewImage':'https://13.design.pages.academy/static/hotel/15.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.828556999999996,'longitude':4.362697,'zoom':16},'isFavorite':false,'isPremium':true,'rating':4.1},{'id':'42b036be-0155-401e-afaa-cc12efcb4064','title':'Nice, cozy, warm big bed apartment','type':'house','price':703,'previewImage':'https://13.design.pages.academy/static/hotel/11.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.835556999999994,'longitude':4.354697,'zoom':16},'isFavorite':false,'isPremium':true,'rating':4.4},{'id':'22c0e878-e960-4925-8f5c-35710fb08fe2','title':'Amazing and Extremely Central Flat','type':'hotel','price':448,'previewImage':'https://13.design.pages.academy/static/hotel/19.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.867557,'longitude':4.339697,'zoom':16},'isFavorite':false,'isPremium':false,'rating':3.7},{'id':'f2333a49-854b-4cdf-9913-fa9523a616a9','title':'House in countryside','type':'house','price':837,'previewImage':'https://13.design.pages.academy/static/hotel/12.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.839557,'longitude':4.346697,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1.1},{'id':'411d8bea-da6c-4439-921f-29f082a9f06c','title':'The Pondhouse - A Magical Place','type':'hotel','price':298,'previewImage':'https://13.design.pages.academy/static/hotel/6.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.844556999999995,'longitude':4.346697,'zoom':16},'isFavorite':false,'isPremium':false,'rating':2.2},{'id':'75c8b13d-a8c3-4afa-8591-bfd39fd247c6','title':'Tile House','type':'house','price':657,'previewImage':'https://13.design.pages.academy/static/hotel/13.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.865556999999995,'longitude':4.371696999999999,'zoom':16},'isFavorite':false,'isPremium':true,'rating':3.5},{'id':'79cf88a4-0699-4644-8243-12b858a9c4f4','title':'Loft Studio in the Central Area','type':'room','price':221,'previewImage':'https://13.design.pages.academy/static/hotel/12.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.869557,'longitude':4.332697,'zoom':16},'isFavorite':false,'isPremium':true,'rating':4.9},{'id':'b5f99874-20ad-45df-b089-d845f0e9e448','title':'House in countryside','type':'room','price':230,'previewImage':'https://13.design.pages.academy/static/hotel/13.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.860557,'longitude':4.376697,'zoom':16},'isFavorite':false,'isPremium':true,'rating':4},{'id':'15766a28-df41-47f4-846e-55296e4a4769','title':'Perfectly located Castro','type':'house','price':652,'previewImage':'https://13.design.pages.academy/static/hotel/20.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.862556999999995,'longitude':4.375697,'zoom':16},'isFavorite':false,'isPremium':false,'rating':4.1},{'id':'d35de079-c332-4ab8-9f11-bce66df089fa','title':'The house among olive ','type':'house','price':798,'previewImage':'https://13.design.pages.academy/static/hotel/8.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.842557,'longitude':4.3536969999999995,'zoom':16},'isFavorite':false,'isPremium':true,'rating':4.9},{'id':'e0d72e2e-32ce-4719-9016-a46f583a4d46','title':'The house among olive ','type':'house','price':774,'previewImage':'https://13.design.pages.academy/static/hotel/6.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.822556999999996,'longitude':4.347697,'zoom':16},'isFavorite':false,'isPremium':false,'rating':3.7},{'id':'163cb7e1-1c76-46f7-8489-615945710433','title':'Wood and stone place','type':'room','price':173,'previewImage':'https://13.design.pages.academy/static/hotel/3.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.867557,'longitude':4.357697,'zoom':16},'isFavorite':false,'isPremium':false,'rating':2.5},{'id':'86d941f6-6ff5-411b-b408-c2ce740f76b7','title':'The Joshua Tree House','type':'house','price':152,'previewImage':'https://13.design.pages.academy/static/hotel/10.jpg','city':{'name':'Brussels','location':{'latitude':50.846557,'longitude':4.351697,'zoom':13}},'location':{'latitude':50.842557,'longitude':4.363696999999999,'zoom':16},'isFavorite':false,'isPremium':false,'rating':3.3},{'id':'d248cab4-2237-4bfc-9520-1d218ed5a85f','title':'Amazing and Extremely Central Flat','type':'apartment','price':292,'previewImage':'https://13.design.pages.academy/static/hotel/6.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.36554,'longitude':4.911976,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2.2},{'id':'44eae572-af54-4ec1-b616-186ac010d053','title':'Wood and stone place','type':'apartment','price':307,'previewImage':'https://13.design.pages.academy/static/hotel/12.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.385540000000006,'longitude':4.902976,'zoom':16},'isFavorite':false,'isPremium':false,'rating':1.7},{'id':'42f63cfc-abde-49ba-8e7c-1265782fad56','title':'Beautiful & luxurious apartment at great location','type':'house','price':365,'previewImage':'https://13.design.pages.academy/static/hotel/2.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.397540000000006,'longitude':4.9099759999999995,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2.1},{'id':'698c2d34-862d-432a-95bc-91aa43b1c1e6','title':'Nice, cozy, warm big bed apartment','type':'apartment','price':454,'previewImage':'https://13.design.pages.academy/static/hotel/6.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.37454,'longitude':4.881976,'zoom':16},'isFavorite':false,'isPremium':false,'rating':3},{'id':'4eec1872-0260-4957-9de9-e8866b2cd3ef','title':'Waterfront with extraordinary view','type':'hotel','price':195,'previewImage':'https://13.design.pages.academy/static/hotel/18.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.367540000000005,'longitude':4.883976,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1.1},{'id':'25991415-0964-4b0f-964a-f63f98ac2a04','title':'The Pondhouse - A Magical Place','type':'apartment','price':437,'previewImage':'https://13.design.pages.academy/static/hotel/10.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.361540000000005,'longitude':4.883976,'zoom':16},'isFavorite':false,'isPremium':true,'rating':3.5},{'id':'cae5d007-5144-4aa0-ba82-0217551500e7','title':'Perfectly located Castro','type':'hotel','price':322,'previewImage':'https://13.design.pages.academy/static/hotel/13.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.35754,'longitude':4.9179759999999995,'zoom':16},'isFavorite':false,'isPremium':false,'rating':3.3},{'id':'12a69063-323c-4c9d-bfdf-595857b787ea','title':'The Pondhouse - A Magical Place','type':'hotel','price':490,'previewImage':'https://13.design.pages.academy/static/hotel/18.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.37854,'longitude':4.894976,'zoom':16},'isFavorite':false,'isPremium':false,'rating':3.3},{'id':'bdc69c15-84c4-4304-869a-d64dc9ee88c1','title':'Amazing and Extremely Central Flat','type':'room','price':142,'previewImage':'https://13.design.pages.academy/static/hotel/18.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.36854,'longitude':4.887976,'zoom':16},'isFavorite':false,'isPremium':false,'rating':2.5},{'id':'84f3dfd3-3715-4f2a-9398-eb0634edd90e','title':'The Pondhouse - A Magical Place','type':'room','price':162,'previewImage':'https://13.design.pages.academy/static/hotel/18.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.370540000000005,'longitude':4.9099759999999995,'zoom':16},'isFavorite':false,'isPremium':true,'rating':3.8},{'id':'e0dbd604-e1c2-4375-b6b2-c578f5f5dbc0','title':'Amazing and Extremely Central Flat','type':'apartment','price':203,'previewImage':'https://13.design.pages.academy/static/hotel/11.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.389540000000004,'longitude':4.883976,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1.7},{'id':'20059b77-a708-4d60-b45f-41b7fc8e5b34','title':'Waterfront with extraordinary view','type':'hotel','price':493,'previewImage':'https://13.design.pages.academy/static/hotel/9.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.35054,'longitude':4.908976,'zoom':16},'isFavorite':false,'isPremium':false,'rating':4.6},{'id':'e347a02b-3ea5-4707-a3bf-69c7e9dd8de6','title':'Loft Studio in the Central Area','type':'room','price':173,'previewImage':'https://13.design.pages.academy/static/hotel/17.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.37154,'longitude':4.889976,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2.6},{'id':'de427268-a490-4cd6-96c4-25b2dfec2752','title':'Beautiful & luxurious apartment at great location','type':'hotel','price':248,'previewImage':'https://13.design.pages.academy/static/hotel/15.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.364540000000005,'longitude':4.9019759999999994,'zoom':16},'isFavorite':false,'isPremium':true,'rating':4.6},{'id':'fa81b898-010a-457d-a514-7f0e2319926c','title':'Wood and stone place','type':'house','price':870,'previewImage':'https://13.design.pages.academy/static/hotel/7.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.36354,'longitude':4.911976,'zoom':16},'isFavorite':false,'isPremium':false,'rating':4.2},{'id':'f3d8fad6-46b9-42b7-a21d-a0d8b4143b6f','title':'Canal View Prinsengracht','type':'apartment','price':490,'previewImage':'https://13.design.pages.academy/static/hotel/1.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.36354,'longitude':4.889976,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2.3},{'id':'f378bb4c-8d67-4893-b383-137680c93ec3','title':'Canal View Prinsengracht','type':'room','price':244,'previewImage':'https://13.design.pages.academy/static/hotel/12.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.388540000000006,'longitude':4.899976,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1.9},{'id':'252edb62-76d7-443b-a8cc-21279800879a','title':'The house among olive ','type':'hotel','price':214,'previewImage':'https://13.design.pages.academy/static/hotel/10.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.36954,'longitude':4.914976,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2.2},{'id':'49891d0c-1550-4f3e-8c8b-eb6303bccdf8','title':'Beautiful & luxurious apartment at great location','type':'room','price':161,'previewImage':'https://13.design.pages.academy/static/hotel/18.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.37554,'longitude':4.9019759999999994,'zoom':16},'isFavorite':false,'isPremium':true,'rating':5},{'id':'e7cee032-82dd-4471-8ceb-b20a30e7003c','title':'Beautiful & luxurious apartment at great location','type':'room','price':165,'previewImage':'https://13.design.pages.academy/static/hotel/11.jpg','city':{'name':'Amsterdam','location':{'latitude':52.37454,'longitude':4.897976,'zoom':13}},'location':{'latitude':52.385540000000006,'longitude':4.886976,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2.4},{'id':'f37d281e-f482-42b5-a8a2-2f0dcbf30bb9','title':'Perfectly located Castro','type':'hotel','price':128,'previewImage':'https://13.design.pages.academy/static/hotel/6.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.528341000000005,'longitude':10.018654000000002,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2},{'id':'eb09f5bd-93de-469c-bfa2-a128b8cdc5f1','title':'Canal View Prinsengracht','type':'apartment','price':367,'previewImage':'https://13.design.pages.academy/static/hotel/20.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.538341,'longitude':9.976654000000002,'zoom':16},'isFavorite':false,'isPremium':false,'rating':4.2},{'id':'edea43a7-630d-4f61-bc49-c985353319d0','title':'Canal View Prinsengracht','type':'house','price':246,'previewImage':'https://13.design.pages.academy/static/hotel/6.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.563341,'longitude':10.019654000000001,'zoom':16},'isFavorite':false,'isPremium':false,'rating':1.6},{'id':'e7c3a18b-ac88-4023-bf57-2f27c13f8990','title':'Loft Studio in the Central Area','type':'house','price':643,'previewImage':'https://13.design.pages.academy/static/hotel/11.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.529341,'longitude':9.975654,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2.2},{'id':'b1d03ab4-ac47-42e1-8732-775851aeae47','title':'Waterfront with extraordinary view','type':'room','price':131,'previewImage':'https://13.design.pages.academy/static/hotel/7.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.565341,'longitude':9.980654000000001,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2.3},{'id':'35a09da9-940a-4505-82dd-e2808008ec7f','title':'Perfectly located Castro','type':'hotel','price':416,'previewImage':'https://13.design.pages.academy/static/hotel/10.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.546341000000005,'longitude':10.022654000000001,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1},{'id':'086ae117-ed56-4a73-a9c6-24c0726f04c1','title':'Canal View Prinsengracht','type':'house','price':749,'previewImage':'https://13.design.pages.academy/static/hotel/17.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.565341000000004,'longitude':9.978654,'zoom':16},'isFavorite':false,'isPremium':false,'rating':4.3},{'id':'a27a1d50-757b-4cfc-be6c-44d95ab1153c','title':'Beautiful & luxurious apartment at great location','type':'house','price':727,'previewImage':'https://13.design.pages.academy/static/hotel/10.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.570341000000006,'longitude':9.975654,'zoom':16},'isFavorite':false,'isPremium':true,'rating':3.1},{'id':'6b3a0723-511d-40f3-9c30-96ef65a2df19','title':'The Pondhouse - A Magical Place','type':'house','price':667,'previewImage':'https://13.design.pages.academy/static/hotel/10.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.558341000000006,'longitude':10.001654,'zoom':16},'isFavorite':false,'isPremium':true,'rating':3.5},{'id':'67a22f72-9cf7-4936-9731-77d5542f3abe','title':'Loft Studio in the Central Area','type':'apartment','price':161,'previewImage':'https://13.design.pages.academy/static/hotel/13.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.573341000000006,'longitude':10.009654000000001,'zoom':16},'isFavorite':false,'isPremium':false,'rating':4.4},{'id':'86daf578-40dd-4c7f-8f93-6388a8d8a2be','title':'Tile House','type':'apartment','price':150,'previewImage':'https://13.design.pages.academy/static/hotel/17.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.573341000000006,'longitude':9.994654,'zoom':16},'isFavorite':false,'isPremium':false,'rating':2},{'id':'e8cf958f-a849-42f1-83cc-46a1add64c40','title':'Beautiful & luxurious apartment at great location','type':'hotel','price':402,'previewImage':'https://13.design.pages.academy/static/hotel/10.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.528341000000005,'longitude':9.982654,'zoom':16},'isFavorite':false,'isPremium':true,'rating':4.6},{'id':'09038918-140e-4e4e-8bb9-fb95444183bb','title':'The house among olive ','type':'room','price':149,'previewImage':'https://13.design.pages.academy/static/hotel/10.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.540341000000005,'longitude':10.025654000000001,'zoom':16},'isFavorite':false,'isPremium':true,'rating':3.2},{'id':'197a696e-488a-40db-b051-76f2431d3cef','title':'Perfectly located Castro','type':'house','price':315,'previewImage':'https://13.design.pages.academy/static/hotel/1.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.555341000000006,'longitude':9.975654,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1.2},{'id':'24fedc35-b59e-4687-a252-cebd42ec190c','title':'Loft Studio in the Central Area','type':'hotel','price':378,'previewImage':'https://13.design.pages.academy/static/hotel/12.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.534341000000005,'longitude':9.998654,'zoom':16},'isFavorite':false,'isPremium':false,'rating':4.5},{'id':'fa4ad916-c0f1-470f-81bc-5eba8c280bc8','title':'Loft Studio in the Central Area','type':'house','price':595,'previewImage':'https://13.design.pages.academy/static/hotel/15.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.563341,'longitude':9.975654,'zoom':16},'isFavorite':false,'isPremium':false,'rating':3.7},{'id':'7b75230b-f8ef-4e15-ab01-66bc01486452','title':'Penthouse, 4-5 rooms + 5 balconies','type':'house','price':495,'previewImage':'https://13.design.pages.academy/static/hotel/11.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.574341000000004,'longitude':10.022654000000001,'zoom':16},'isFavorite':false,'isPremium':true,'rating':5},{'id':'e7855226-4010-4f5f-8652-b9d15c3fe983','title':'The house among olive ','type':'apartment','price':470,'previewImage':'https://13.design.pages.academy/static/hotel/13.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.573341000000006,'longitude':10.025654000000001,'zoom':16},'isFavorite':false,'isPremium':true,'rating':4},{'id':'7b9b3720-8d5b-485e-abc3-dca4ae3bfd7b','title':'The Pondhouse - A Magical Place','type':'room','price':204,'previewImage':'https://13.design.pages.academy/static/hotel/2.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.550341,'longitude':9.980654000000001,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1.3},{'id':'1ef70cf0-68c2-4437-9870-5497d6202f93','title':'Penthouse, 4-5 rooms + 5 balconies','type':'room','price':263,'previewImage':'https://13.design.pages.academy/static/hotel/7.jpg','city':{'name':'Hamburg','location':{'latitude':53.550341,'longitude':10.000654,'zoom':13}},'location':{'latitude':53.558341000000006,'longitude':9.999654000000001,'zoom':16},'isFavorite':false,'isPremium':false,'rating':4.7},{'id':'dfffca4a-070c-4b53-a31f-bcd702f28cbf','title':'The Joshua Tree House','type':'room','price':199,'previewImage':'https://13.design.pages.academy/static/hotel/1.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.236402000000005,'longitude':6.784314,'zoom':16},'isFavorite':false,'isPremium':false,'rating':1.1},{'id':'dc3f6388-4f97-4018-b758-5f36f0a4de29','title':'Penthouse, 4-5 rooms + 5 balconies','type':'apartment','price':332,'previewImage':'https://13.design.pages.academy/static/hotel/20.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.210402,'longitude':6.798314,'zoom':16},'isFavorite':false,'isPremium':false,'rating':1.6},{'id':'850dcaeb-6eb7-4f76-9b83-41746b41c318','title':'The Pondhouse - A Magical Place','type':'room','price':252,'previewImage':'https://13.design.pages.academy/static/hotel/8.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.211402,'longitude':6.756314000000001,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1.4},{'id':'26e2b19e-4a3e-4223-96d0-e0efeb3e95d3','title':'Nice, cozy, warm big bed apartment','type':'hotel','price':153,'previewImage':'https://13.design.pages.academy/static/hotel/6.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.228402,'longitude':6.784314,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2.6},{'id':'c1419eec-5005-4c17-af91-aa42a2b884eb','title':'Wood and stone place','type':'room','price':232,'previewImage':'https://13.design.pages.academy/static/hotel/16.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.204402,'longitude':6.7773140000000005,'zoom':16},'isFavorite':false,'isPremium':false,'rating':2},{'id':'bba30aeb-2469-4e3c-a540-7bafb33b55c6','title':'Perfectly located Castro','type':'hotel','price':281,'previewImage':'https://13.design.pages.academy/static/hotel/8.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.217402,'longitude':6.7693140000000005,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2.2},{'id':'e697aa10-a09d-4261-93b7-dc38944fabac','title':'The house among olive ','type':'apartment','price':222,'previewImage':'https://13.design.pages.academy/static/hotel/2.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.241402,'longitude':6.782314,'zoom':16},'isFavorite':false,'isPremium':true,'rating':2.9},{'id':'5e700a79-2311-4054-adb0-caf6f11470db','title':'Penthouse, 4-5 rooms + 5 balconies','type':'apartment','price':399,'previewImage':'https://13.design.pages.academy/static/hotel/16.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.235402,'longitude':6.800314,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1.5},{'id':'815fde23-6b34-443a-a276-129c2beb80a8','title':'The Pondhouse - A Magical Place','type':'hotel','price':430,'previewImage':'https://13.design.pages.academy/static/hotel/2.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.237402,'longitude':6.779314,'zoom':16},'isFavorite':false,'isPremium':false,'rating':4.7},{'id':'70e4c9fd-1ec3-4570-8c13-9a714d8a7c66','title':'The Pondhouse - A Magical Place','type':'hotel','price':324,'previewImage':'https://13.design.pages.academy/static/hotel/14.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.237402,'longitude':6.797314,'zoom':16},'isFavorite':false,'isPremium':false,'rating':5},{'id':'79ee90ee-4bff-4aa4-8953-dc8251b621f8','title':'Loft Studio in the Central Area','type':'room','price':239,'previewImage':'https://13.design.pages.academy/static/hotel/16.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.222402,'longitude':6.786314,'zoom':16},'isFavorite':false,'isPremium':false,'rating':1.2},{'id':'7d67c7e2-0782-4499-bcc6-1deb78f51ebc','title':'The Pondhouse - A Magical Place','type':'house','price':820,'previewImage':'https://13.design.pages.academy/static/hotel/3.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.232402,'longitude':6.800314,'zoom':16},'isFavorite':false,'isPremium':false,'rating':3.5},{'id':'e7c02c9c-5489-4b0b-8740-369d0a6b9ada','title':'Loft Studio in the Central Area','type':'room','price':229,'previewImage':'https://13.design.pages.academy/static/hotel/4.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.216402,'longitude':6.758314,'zoom':16},'isFavorite':false,'isPremium':true,'rating':3.9},{'id':'a97780bb-0ff8-4962-a074-3745fb000428','title':'Nice, cozy, warm big bed apartment','type':'apartment','price':191,'previewImage':'https://13.design.pages.academy/static/hotel/20.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.205402,'longitude':6.7613140000000005,'zoom':16},'isFavorite':false,'isPremium':false,'rating':3.1},{'id':'7fcefc06-2969-4188-8d9c-496c9b2c4177','title':'Canal View Prinsengracht','type':'room','price':224,'previewImage':'https://13.design.pages.academy/static/hotel/7.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.248402000000006,'longitude':6.763314,'zoom':16},'isFavorite':false,'isPremium':true,'rating':3.8},{'id':'6b9dfe86-17d3-4c0f-b2a5-4780066e3e53','title':'The house among olive ','type':'apartment','price':329,'previewImage':'https://13.design.pages.academy/static/hotel/7.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.250402,'longitude':6.7853140000000005,'zoom':16},'isFavorite':false,'isPremium':false,'rating':1.3},{'id':'f0abb3a7-3906-40d4-ada6-405497a0578f','title':'Beautiful & luxurious apartment at great location','type':'house','price':741,'previewImage':'https://13.design.pages.academy/static/hotel/9.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.239402000000005,'longitude':6.756314000000001,'zoom':16},'isFavorite':false,'isPremium':false,'rating':3.5},{'id':'58f643ce-fc3a-4a20-b87e-b0436af069b2','title':'Beautiful & luxurious apartment at great location','type':'hotel','price':108,'previewImage':'https://13.design.pages.academy/static/hotel/3.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.243402,'longitude':6.791314,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1.8},{'id':'6b51faee-56cf-41bd-9a8d-ba1fa381816a','title':'The Pondhouse - A Magical Place','type':'house','price':829,'previewImage':'https://13.design.pages.academy/static/hotel/16.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.214402,'longitude':6.764314000000001,'zoom':16},'isFavorite':false,'isPremium':true,'rating':1},{'id':'10106d2c-7500-45de-bec7-abb8208e5c29','title':'Wood and stone place','type':'apartment','price':364,'previewImage':'https://13.design.pages.academy/static/hotel/1.jpg','city':{'name':'Dusseldorf','location':{'latitude':51.225402,'longitude':6.776314,'zoom':13}},'location':{'latitude':51.225402,'longitude':6.784314,'zoom':16},'isFavorite':false,'isPremium':false,'rating':1.4}];
+
+export default mocks;
